Move expand btn, add icons, change list view
