+++ frontend/src/components/WorksheetExpandedList.js
@@ -1,6 +1,7 @@
 import React, { useState, useMemo } from 'react';
 import { Col, Row, Button } from 'react-bootstrap';
 import { FaCompressAlt } from 'react-icons/fa';
+import { BsArrowLeft } from 'react-icons/bs';
 import SearchResults from './SearchResults';
 import FBReactSelect from './FBReactSelect';
 import SeasonReactSelect from './SeasonReactSelect';
@@ -69,11 +70,33 @@ const WorksheetExpandedList = ({
   return (
     <div className={styles.container}>
       <Row className="mx-auto">
+        {/* Worksheet courses in search results format */}
+        <Col md={9} className="pr-3 pl-0">
+          <div className={styles.search_results}>
+            <SearchResults
+              data={WorksheetData}
+              showModal={showModal}
+              expanded={cur_expand !== 'list'}
+              isLoggedIn
+              isList={isList}
+              setView={setView}
+              num_fb={num_fb}
+            />
+          </div>
+        </Col>
         {/* Season and FB friends dropdown */}
         <Col md={3} className="p-0">
-          <StyledExpandLink variant='link' className='mb-2' onClick={() => setCurExpand('none')}>Go to calendar view</StyledExpandLink>
           <SurfaceComponent layer={0} className={`${styles.select_col} p-2`}>
             <Row className="mx-auto">
+              <StyledExpandLink
+                variant="link"
+                className="py-0"
+                onClick={() => setCurExpand('none')}
+              >
+                <BsArrowLeft size={24} /> Go to calendar view
+              </StyledExpandLink>
+            </Row>
+            <Row className="mx-auto mt-2">
               <div
                 className={`${select_styles.select_container} ${select_styles.hover_effect}`}
               >
@@ -84,7 +107,7 @@ const WorksheetExpandedList = ({
                 />
               </div>
             </Row>
-            <Row className="mx-auto my-2">
+            <Row className="mx-auto mt-2">
               <div
                 className={
                   select_styles.select_container +
@@ -98,7 +121,7 @@ const WorksheetExpandedList = ({
                 />
               </div>
             </Row>
-            <Row className="mx-auto">
+            <Row className="mx-auto mt-2">
               <SortbyReactSelect setOrdering={setOrdering} />
             </Row>
             {/* <StyledExpandBtn
@@ -114,20 +137,6 @@ const WorksheetExpandedList = ({
               />
             </StyledExpandBtn> */}
           </SurfaceComponent>
-        </Col>
-        {/* Worksheet courses in search results format */}
-        <Col md={9} className="pr-0 pl-3">
-          <div className={styles.search_results}>
-            <SearchResults
-              data={WorksheetData}
-              showModal={showModal}
-              expanded={cur_expand !== 'list'}
-              isLoggedIn
-              isList={isList}
-              setView={setView}
-              num_fb={num_fb}
-            />
-          </div>
         </Col>
       </Row>
     </div>
